tbc: split TBCInstr.hpp and hoist VM core into TBCVm.hpp
Preparation for the copy-and-patch JIT execution mode:

- TBCInstr.hpp: freestanding Instr/kNoReg/packedOffset* definitions
  (no <string>/<vector>), includable by the future stencil TU.
- TBCVm.hpp/.cpp: VMContext, pushFrame, doExtCall, doIndirectCall
  hoisted out of TBCInterpreter.cpp's anonymous namespace so stitched
  code's plain-ABI helpers can share the exact frame logic. VMContext's
  hot members (sp/stackEnd/prog) now come first: their offsets become
  part of the stencil ABI. Adds a pendingException slot for the JIT's
  exception bridging.

No behavior change; all 349 tests pass.
